+++ 4. Domain/Placica.Core.Library/Contracts/DomainServices/IDomainService.cs
@@ -9,10 +9,10 @@ namespace Placica.Core.Library.Contracts.DomainServices
         where T : class, IEntity
     {
         Task<IEnumerable<T>> GetAll();
-        Task<T> Get(int id);
+        Task<T> Get(long id);
         Task<T> Add(T entity);
         Task<T> Update(T entity);
-        Task<T> Delete(int id);
+        Task<T> Delete(long id);
         IQueryable<T> GetAllQueryable();
     }
 }
